+++ src/resources/views/site/element/head.blade.php
@@ -2,3 +2,13 @@
 <style>
     {!! !empty($config[$theme.'_css']) ? $config[$theme.'_css'] : ''  !!}
 </style>
+
+@if(config('services.recaptcha.enable'))
+    <script
+        src="https://www.google.com/recaptcha/api.js?render={{ config('services.recaptcha.site_key') }}"></script>
+    <script>
+        function onSubmit(token) {
+            document.getElementsByClassName("recaptcha")[0].submit();
+        }
+    </script>
+@endif