+++ src/resources/views/site/layout/tool.blade.php
@@ -44,10 +44,7 @@
 
     </style>
 
-    {!! !empty($config['code_header']) ? $config['code_header'] : ''  !!}
-    <style>
-        {!! !empty($config[$theme.'_css']) ? $config[$theme.'_css'] : ''  !!}
-    </style>
+    @include('site.element.head')
 </head>
 
 <body>