+++ src/resources/views/site/member/login.blade.php
@@ -6,7 +6,7 @@
             @include("site.member.login_social")
 
             @if(($config['login_basic_app_status'] ?? '') == 'on')
-                <form method="post" action="{{ base_url('member/login') }}">
+                <form method="post" class="recaptcha" action="{{ base_url('member/login') }}">
                     @csrf
                     <div class="loginbox-textbox">
                         <label class="control-label">Email</label>
@@ -24,7 +24,20 @@
                     </div>
 
                     <div class="loginbox-submit">
-                        <input type="submit" class="btn btn-primary btn-block" value="Login">
+                        @if(config('services.recaptcha.enable'))
+                            <button class="g-recaptcha btn btn-primary btn-block"
+                                    data-sitekey="{{ config('services.recaptcha.site_key') }}"
+                                    data-callback='onSubmit'
+                                    data-action='submit'>
+                                <i class="fa fa-sign-in"></i>
+                                Login
+                            </button>
+                        @else
+                            <button class="btn btn-primary btn-block">
+                                <i class="fa fa-sign-in"></i>
+                                Login
+                            </button>
+                        @endif
                     </div>
 
                     <div class="loginbox-textbox">