+++ CHANGES.md
@@ -4,3 +4,9 @@ Laravel Wordpress Cms Changelog
 ## Update Release Code
 
 * 2021-03-09: update code system switch case ckeditor & summernote
+* 2021-03-25: add google recaptcha for form login, register in admin and frontend
+	- file .env update
+		+ GOOGLE_RECAPTCHA_ENABLED=false
+		+ GOOGLE_RECAPTCHA_KEY=
+		+ GOOGLE_RECAPTCHA_SECRET=
+	- how to on/off google recaptcha GOOGLE_RECAPTCHA_ENABLED=true/false